// src/generated/data-contracts.ts
export enum CurrencyEnum {
  USD = 'USD',
  ZAR = 'ZAR',
  AED = 'AED',
}

/** The country in which the business was incorporated */
export enum IncorporationCountryEnum {
  EG = 'EG',
  JO = 'JO',
  SA = 'SA',
  AE = 'AE',
  ES = 'ES',
  ZA = 'ZA',
}

/**
 * The credit score associated to the customer
 * @example 1
 */
export enum CreditScoreEnum {
  Value1 = 1,
  Value2 = 2,
  Value3 = 3,
  Value4 = 4,
  Value5 = 5,
  Value6 = 6,
  Value7 = 7,
  Value8 = 8,
}

export interface MetaData {
  /** Number of items */
  itemsCount?: number;
  /** Limit of items per page */
  limit?: number;
  /** Current page number */
  page?: number;
}

/**
 * Internal risk associated with the customer
 * @example "LOW"
 */
export enum InternalRisk {
  VERY_HIGH = 'VERY_HIGH',
  HIGH = 'HIGH',
  NEUTRAL = 'NEUTRAL',
  LOW = 'LOW',
}

export enum OrderByEnum {
  Asc = 'asc',
  Desc = 'desc',
}
// ...
export interface NewCompany {
  /**
   * Company Legal Name
   * @example "My Company Inc"
   */
  companyLegalName: string;
  /** The country in which the business was incorporated */
  incorporationCountry: IncorporationCountryEnum;
  /**
   * The date on which the company was incorporated. A date in ISO-8601 format (YYYY-MM-DD)
   * @example "2023-01-01T00:00:00.000Z"
   */
  incorporationDate: string;
  /**
   * Type of legal entity
   * @example "LLC"
   */
  companyType: string;
  /**
   * Industry of the company
   * @example "RETAIL"
   */
  industry?: string;
  /**
   * Indicates if the company is a partner
   * @example false
   */
  isPartner?: boolean;
  /**
   * The partner to which this company belongs
   * @example "Partner Company Inc EG"
   */
  partnerCompanyIdentifier?: string;
  /**
   * Contact email for company
   * @example "user@example.com"
   */
  email: string;
  /**
   * Contact phone for customer
   * @example "+20961234567"
   */
  phone: string;
  /**
   * Website for this customer
   * @example "https://flow48.com"
   */
  website?: string;
  /**
   * Name of the contact associated with this company
   * @example "John Doe"
   */
  contactName: string;
  /**
   * The owner is the Sales team member managing the relationship with the client
   * @example "123e4567-e89b-12d3-a456-426614174000"
   */
  loanOfficerUserId: string;
  /**
   * The credit limit associated with the customer
   * @example 100000
   */
  creditLimit: number;
  /**
   * Credit limit associated with the IBF product
   * @example 50000
   */
  ibfCreditLimit?: number;
  /** Currency */
  currency?: CurrencyEnum;
  /**
   * Credit limit associated with the IBF product
   * @example 50000
   */
  rbfCreditLimit?: number;
  /** The credit score associated to the customer */
  creditScore: CreditScoreEnum;
  /**
   * Date of which the customer needs next risk review
   * @example "2029-01-01T00:00:00.000Z"
   */
  nextReviewDate: string;
  /**
   * Indicates if the company has compliance approved
   * @example true
   */
  isComplianceApproved: boolean;
  /**
   * Next date when compliance is due
   * @example "2029-01-01T00:00:00.000Z"
   */
  nextComplianceReviewDate: string;
  /**
   * Name of the company bank
   * @example "Dark Empire Bank"
   */
  bankName: string;
  /**
   * IBAN code of the company bank
   * @example "EG800002000156789012345180002"
   */
  iban: string;
  /**
   * SWIFT code for the company bank
   * @example "BBBBJO33YYY"
   */
  swift: string;
  /**
   * Address of company's bank
   * @example "Coruscant street 6"
   */
  bankAddress: string;
  /**
   * The bank account number
   * @example 1234567
   */
  accountNumber?: string;
  /**
   * The bank branch code
   * @example 123456
   */
  branchCode?: string;
  /** Internal risk associated with the customer */
  internalRisk?: InternalRisk;
}

export type CompanyResponse = (NewCompany & object) & {
  /**
   * Unique company identifier
   * @example "My Company Inc EG"
   */
  companyIdentifier?: string;
  /** @example "Direct" */
  companySourceType?: CompanyResponseCompanySourceTypeEnum;
  /**
   * The difference between Credit Limit and Utilized Credit
   * @example 100000
   */
  limitGap?: number;
  /**
   * The credit utilized by the customer, the value of the credits - total principal repayments
   * @example 100000
   */
  utilizedCredit?: number;
  /**
   * The IBF credit utilized by the customer
   * @example 100000
   */
  utilizedIbfCredit?: number;
  /**
   * The IBRBFF credit utilized by the customer
   * @example 100000
   */
  utilizedRbfCredit?: number;
  /**
   * Date and time when company was added in system
   * @example "2023-05-06T23:44:08.209Z"
   */
  createdAt?: string;
  /**
   * Date and time when company was last updated
   * @example "2023-05-06T23:44:08.209Z"
   */
  updatedAt?: string;
  /**
   * Who the company was last updated by
   * @example "123e4567-e89b-12d3-a456-426614174000"
   */
  updatedByUserId?: string;
  loanOfficerUser?: LoanOfficerUser;
  /**
   * Additional company attributes
   * @example {"attributeName":"attribute Value"}
   */
  attributes?: object;
};

export interface SalesforceUser {
  /** @example "0052v00000B8Z3AAAV" */
  Id?: string;
  /** @example "John Doe" */
  Name?: string;
  /** @example "sales@flow48.com" */
  Email?: string;
}

export interface Account {
  /** @example "0012v00002J5Z3AAAV" */
  Id?: string;
  /** @example "My Company Inc EG" */
  Company_Identifier__c?: string;
  /** @example "My Company Inc EG" */
  Name?: string;
  /** @example "https://flow48.com" */
  Website?: string;
  /** The country in which the business was incorporated */
  IncorporationCountry__c?: IncorporationCountryEnum;
  /** @example "LLC" */
  BusinessType__c?: string;
  /** @example "RETAIL" */
  Industry?: string;
  Parent?: Account;
  /** @example "contact@email.com" */
  Email__c?: string;
  /** @example "+20961234567" */
  Phone?: string;
  /** @example "John Doe" */
  ContactName__c?: string;
  /** @example "0052v00000B8Z3AAAV" */
  OwnerId?: string;
  Owner?: SalesforceUser;
  /** @example 100000 */
  CreditLimit__c?: number;
  /** @example 1 */
  CreditScore__c?: number;
  /** @example true */
  ComplianceApproved__c?: boolean;
  /** @example "Dark Empire Bank" */
  BankName__c?: string;
  /** @example "BBBBJO33YYY" */
  SWIFT__c?: string;
  /** @example "EG800002000156789012345180002" */
  IBAN__c?: string;
  /** @example "Coruscant street 6" */
  BankAddress__c?: string;
}
// ...
export type HalLink = Record<
  string,
  {
    href?: string;
  }
>;

export interface CompanyLinks {
  getCompany?: {
    /** @example "/companies/{companyId}" */
    href?: string;
  };
  getCompanies?: {
    /** @example "/companies" */
    href?: string;
  };
  createCompany?: {
    /** @example "/companies" */
    href?: string;
  };
  createLoan?: {
    /** @example "/loans" */
    href?: string;
  };
  getLoans?: {
    /** @example "/loans" */
    href?: string;
  };
  getCompanyLoans?: {
    /** @example "/companies/{companyId}/loans" */
    href?: string;
  };
  getCompanyActionLogs?: {
    /** @example "/companies/{companyId}/actionLogs" */
    href?: string;
  };
}

export interface CompaniesLinks {
  getCompanies?: {
    /** @example "/companies" */
    href?: string;
  };
  createCompany?: {
    /** @example "/companies" */
    href?: string;
  };
  createLoan?: {
    /** @example "/loans" */
    href?: string;
  };
  getLoans?: {
    /** @example "/loans" */
    href?: string;
  };
}

export interface FormFieldsResponse {
  fieldName?: string[];
}

export interface FormFieldsValidationResponse {
  /** @example "Some error" */
  fieldName?: any;
}

export interface LoanOfficerUser {
  /** @example "123e4567-e89b-12d3-a456-426614174000" */
  user_id?: string;
  /** @example "example@example.com" */
  email?: string;
}

export interface ActionLog {
  /** @example 1 */
  logId?: string;
  /**
   * Company identifier
   * @example "My Company Inc EG"
   */
  companyIdentifier?: string;
  /**
   * The ID of the entity for which the action was performed
   * @example "7j6ia3F7pcgmYox"
   */
  entityId?: string;
  /**
   * Entity type
   * @example "Company"
   */
  entityType?: string;
  /**
   * Action log data
   * @example "Company is created"
   */
  data?: string;
  /**
   * Date and time when log was added in system
   * @example "2023-05-06T23:44:08.000Z"
   */
  created_at?: string;
  loanOfficerUser?: LoanOfficerUser;
}
// ...
export enum CompanyResponseCompanySourceTypeEnum {
  Partner = 'Partner',
  Direct = 'Direct',
  PartnerCustomer = 'Partner customer',
}
// ...
export interface GetCompaniesParams {
  /** In which order will the data be sorted (ascending or descending) */
  orderBy?: OrderByEnum;
  /** Search items by partial matching of search param */
  search?: string;
  /** Number of items in the response */
  limit?: number;
  /** Page number */
  page?: number;
  /** Fields that can be included */
  include?: IncludeEnum[];
}

export enum IncludeEnum {
  IncorporationDate = 'incorporationDate',
  CompanyType = 'companyType',
  Industry = 'industry',
  IsPartner = 'isPartner',
  PartnerCompany = 'partnerCompany',
  Email = 'email',
  Phone = 'phone',
  Website = 'website',
  ContactName = 'contactName',
  LoanOfficerUserId = 'loanOfficerUserId',
  IbfCreditLimit = 'ibfCreditLimit',
  RbfCreditLimit = 'rbfCreditLimit',
  CreditLimit = 'creditLimit',
  CreditScore = 'creditScore',
  NextReviewDate = 'nextReviewDate',
  IsComplianceApproved = 'isComplianceApproved',
  NextComplianceReviewDate = 'nextComplianceReviewDate',
  BankName = 'bankName',
  Iban = 'iban',
  Swift = 'swift',
  BankAddress = 'bankAddress',
  CreatedAt = 'createdAt',
  UpdatedAt = 'updatedAt',
  UpdatedByUserId = 'updatedByUserId',
}

export interface GetCompaniesData {
  companies?: CompanyResponse[];
}
// ...
export type CreateCompanyData = (CompanyResponse & object) & {
  _links?: HalLink[];
};

export type GetCompanyData = (CompanyResponse & object) & {
  _links?: CompanyLinks[];
};

export interface GetCompanyActionLogParams {
  /** In which order will the data be sorted (ascending or descending) */
  orderBy?: OrderByEnum;
  /** Field by which the records will be sorted */
  sortBy?: SortByEnum;
  /** Number of items in the response */
  limit?: number;
  /** Page number */
  page?: number;
  /**
   * Company identifier
   * @example "My Company Inc EG"
   */
  companyId: string;
}

/** Field by which the records will be sorted */
export enum SortByEnum {
  CreatedAt = 'createdAt',
}

export type GetCompanyActionLogData = ActionLog & object;

/** Field by which the records will be sorted */
export enum GetCompanyActionLogParams1SortByEnum {
  CreatedAt = 'createdAt',
}

export interface GetCompanyLoansParams {
  /** In which order will the data be sorted (ascending or descending) */
  orderBy?: OrderByEnum;
  /** Number of items in the response */
  limit?: number;
  /** Page number */
  page?: number;
  /** Company field by which the records will be sorted */
  sortBy?: SortByEnum1;
  /** Loan fields that can be included */
  include?: IncludeEnum1[];
  /**
   * Company identifier
   * @example "My Company Inc EG"
   */
  companyId: string;
}

/** Company field by which the records will be sorted */
export enum SortByEnum1 {
  CompanyIdentifier = 'companyIdentifier',
  LoanOfficerUserId = 'loanOfficerUserId',
  LoanType = 'loanType',
  GracePeriodDays = 'gracePeriodDays',
  Tenor = 'tenor',
  PenaltyPercent = 'penaltyPercent',
  PenaltyType = 'penaltyType',
  MonthlyRate = 'monthlyRate',
  LoanStatus = 'loanStatus',
  DisbursementDate = 'disbursementDate',
  MaturityDate = 'maturityDate',
  PayoffDate = 'payoffDate',
  LoanApprovalDate = 'loanApprovalDate',
  TimesDelinquent = 'timesDelinquent',
  PrincipalAmount = 'principalAmount',
  InterestAmount = 'interestAmount',
  TotalDisbursedAmount = 'totalDisbursedAmount',
  PaidPrincipalAmount = 'paidPrincipalAmount',
  PaidInterestAmount = 'paidInterestAmount',
  PaidPenaltiesAmount = 'paidPenaltiesAmount',
  TotalPenaltyAmount = 'totalPenaltyAmount',
  CreatedAt = 'createdAt',
}

export enum IncludeEnum1 {
  SalesforceOpportunityId = 'salesforceOpportunityId',
  LoanOfficerUserId = 'loanOfficerUserId',
  GracePeriodDays = 'gracePeriodDays',
  PenaltyPercent = 'penaltyPercent',
  PenaltyType = 'penaltyType',
  LoanStatus = 'loanStatus',
  DisbursementDate = 'disbursementDate',
  MaturityDate = 'maturityDate',
  PayoffDate = 'payoffDate',
  LoanApprovalDate = 'loanApprovalDate',
  TimesDelinquent = 'timesDelinquent',
  TotalDisbursedAmount = 'totalDisbursedAmount',
  PaidPrincipalAmount = 'paidPrincipalAmount',
  PaidInterestAmount = 'paidInterestAmount',
  PaidPenaltiesAmount = 'paidPenaltiesAmount',
  TotalPenaltyAmount = 'totalPenaltyAmount',
  CreatedAt = 'createdAt',
}

export interface GetCompanyLoansData {
  loans?: LoanResponse[];
}
// ...
export type GetAdditionCompanyFormFieldsData = FormFieldsResponse;

export interface ValidateAdditionCompanyFormFieldsData {
  selectFields?: {
    fieldName?: string[];
  };
}

export type GetCompanyBySalesforceIdData = (CompanyResponse & object) & {
  _links?: HalLink[];
};

export type PutCompanyBySalesforceIdData = (CompanyResponse & object) & {
  _links?: HalLink[];
};